Possible Update to hysplit_dispersion.R
Unfortunately, I am relatively new to R so this edit may be incorrect. I am trying to use SplitR on a Windows 10 machine but the script is having trouble writing the "dispersion.csv" file. I'm not certain, but I think it has to do with the way your slashes were set up. I will keep playing with it on my end. Thank you VERY much for this package! It's brilliant!