Working correction to path construction
I copied lines 807-815 into a separate script and got it to work like this. I do believe it was an issue with your slashes. Also, I used a short piece of code for the "get_os" function from a website that didn't actually include an if statement for Windows machines. I'm not sure if you made your own custom "get_os" function or not.